+++ test/hessenberg.jl
@@ -5,9 +5,6 @@ module TestHessenberg
 using Test, LinearAlgebra, Random
 
 const BASE_TEST_PATH = joinpath(Sys.BINDIR, "..", "share", "julia", "test")
-isdefined(Main, :Furlongs) || @eval Main include(joinpath($(BASE_TEST_PATH), "testhelpers", "Furlongs.jl"))
-using .Main.Furlongs
-
 isdefined(Main, :SizedArrays) || @eval Main include(joinpath($(BASE_TEST_PATH), "testhelpers", "SizedArrays.jl"))
 using .Main.SizedArrays
 
@@ -68,29 +65,20 @@ let n = 10
         @test Array(Hc + H) == Array(Hc) + Array(H)
         @test Array(Hc - H) == Array(Hc) - Array(H)
         @testset "Preserve UpperHessenberg shape (issue #39388)" begin
-            for H = (UpperHessenberg(Areal), UpperHessenberg(Furlong.(Areal)))
-                if eltype(H) <: Furlong
-                    A = Furlong.(rand(n,n))
-                    d = Furlong.(rand(n))
-                    dl = Furlong.(rand(n-1))
-                    du = Furlong.(rand(n-1))
-                    us = Furlong(1)*I
-                else
-                    A = rand(n,n)
-                    d = rand(n)
-                    dl = rand(n-1)
-                    du = rand(n-1)
-                    us = 1*I
-                end
-                @testset "$op" for op = (+,-)
-                    for x = (us, Diagonal(d), Bidiagonal(d,dl,:U), Bidiagonal(d,dl,:L),
-                             Tridiagonal(dl,d,du), SymTridiagonal(d,dl),
-                             UpperTriangular(A), UnitUpperTriangular(A))
-                        @test op(H,x) == op(Array(H),x)
-                        @test op(x,H) == op(x,Array(H))
-                        @test op(H,x) isa UpperHessenberg
-                        @test op(x,H) isa UpperHessenberg
-                    end
+            H = UpperHessenberg(Areal)
+            A = rand(n,n)
+            d = rand(n)
+            dl = rand(n-1)
+            du = rand(n-1)
+            us = 1*I
+            @testset "$op" for op = (+,-)
+                for x = (us, Diagonal(d), Bidiagonal(d,dl,:U), Bidiagonal(d,dl,:L),
+                            Tridiagonal(dl,d,du), SymTridiagonal(d,dl),
+                            UpperTriangular(A), UnitUpperTriangular(A))
+                    @test op(H,x) == op(Array(H),x)
+                    @test op(x,H) == op(x,Array(H))
+                    @test op(H,x) isa UpperHessenberg
+                    @test op(x,H) isa UpperHessenberg
                 end
             end
             H = UpperHessenberg(Areal)
@@ -102,8 +90,8 @@ let n = 10
                             UpperTriangular(A), UnitUpperTriangular(A))
                     @test (H*x)::UpperHessenberg ≈ Array(H)*x
                     @test (x*H)::UpperHessenberg ≈ x*Array(H)
-                    @test H/x ≈ Array(H)/x# broken = eltype(H) <: Furlong && x isa UpperTriangular
-                    @test x\H ≈ x\Array(H)# broken = eltype(H) <: Furlong && x isa UpperTriangular
+                    @test H/x ≈ Array(H)/x
+                    @test x\H ≈ x\Array(H)
                     @test H/x isa UpperHessenberg
                     @test x\H isa UpperHessenberg
                 end
@@ -112,23 +100,6 @@ let n = 10
                 @test x*H == x*Array(H)
                 @test H/x == Array(H)/x
                 @test x\H == x\Array(H)
-            end
-            H = UpperHessenberg(Furlong.(Areal))
-            for A in (A, Furlong.(A))
-                @testset "Multiplication/division Furlong" begin
-                    for x = (5, 5I, Diagonal(d), Bidiagonal(d,dl,:U),
-                                UpperTriangular(A), UnitUpperTriangular(A))
-                        @test map(x -> x.val, (H*x)::UpperHessenberg) ≈ map(x -> x.val, Array(H)*x)
-                        @test map(x -> x.val, (x*H)::UpperHessenberg) ≈ map(x -> x.val, x*Array(H))
-                        @test map(x -> x.val, (H/x)::UpperHessenberg) ≈ map(x -> x.val, Array(H)/x)
-                        @test map(x -> x.val, (x\H)::UpperHessenberg) ≈ map(x -> x.val, x\Array(H))
-                    end
-                    x = Bidiagonal(d, dl, :L)
-                    @test H*x == Array(H)*x
-                    @test x*H == x*Array(H)
-                    @test H/x == Array(H)/x
-                    @test x\H == x\Array(H)
-                end
             end
         end
     end
