+++ test/testgroups
@@ -1,4 +1,5 @@
 triangular
+triangular2
 addmul
 bidiag
 bitarray
@@ -6,6 +7,7 @@ matmul
 dense
 symmetric
 diagonal
+unitful
 special
 qr
 cholesky
